HH-188149 move DataSourceContextUnsafe to nab-data-source module
