+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/transaction/DataSourceContext.java
@@ -2,6 +2,7 @@ package ru.hh.nab.hibernate.transaction;
 
 import java.util.function.Supplier;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
+import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.datasource.DataSourceType;
 
 public class DataSourceContext {
