+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceAspect.java
@@ -9,6 +9,7 @@ import static org.springframework.transaction.TransactionDefinition.PROPAGATION_
 import static org.springframework.transaction.TransactionDefinition.PROPAGATION_REQUIRED;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 import org.springframework.transaction.support.TransactionTemplate;
+import ru.hh.nab.datasource.DataSourceContextUnsafe;
 
 @Aspect
 public class ExecuteOnDataSourceAspect {
